[FIX] website_forum: missing parenthesis for function
before this commit, on clicking on Go To Website
smart button in forum is not redirecting to the
corresponding forum due to missing parenthesis
to the function.

after this commit, on clicking the Go To Website
smart button, user is redirected to the forum.

introduced in: https://github.com/odoo/odoo/commit/36c734ab37fdf1d6a245e0e29cfefe3cc46c6d92

closes odoo/odoo#121373

X-original-commit: 2f22565102d1f472fb700f95caf4d135cd2c34f0
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>